- Fixed several bugs. - Fixed bugs with the DXGI adapter. Added a new form to create the swap chain and d3dDevice. - Added support for fullscreen mode.

+++ GraphicsEngine/GraphicsEngine/GraphicsEngine/VertexShader.cpp
@@ -34,7 +34,7 @@ void VertexShader::Initialize(ID3D11Device* d3dDevice, const void* shaderByteCod
 	ThrowIfFailed(
 		d3dDevice->CreateInputLayout(
 			vertexDesc.data(),
-			vertexDesc.size(),
+			static_cast<UINT>(vertexDesc.size()),
 			shaderByteCode,
 			byteCodeLength,
 			m_inputLayout.GetAddressOf()
